# cmd/target-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ownership verification and authorization scope gate for the Scantinel security-scanning platform. The POST /internal/v1/scope/check endpoint is the hard gate that scan-service and agent-service MUST consult before dispatching any intrusive task.",
        "title": "target-service API",
        "termsOfService": "https://scantinel.ai/terms",
        "contact": {
            "name": "Scantinel Support",
            "url": "https://scantinel.ai/contact",
            "email": "security@scantinel.ai"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/internal/v1/scope/check": {
            "post": {
                "description": "Cluster-only. Determines whether a given host or IP is authorized for an intrusive scan phase on behalf of the target owner. Scan-service and agent-service MUST call this before dispatching any intrusive task.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Authorization scope gate (INTERNAL)",
                "parameters": [
                    {
                        "description": "Scope check payload: {target_uid, host|ip, phase}",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ScopeCheckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ScopeCheckResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/internal/v1/targets/{target_uid}/assets": {
            "post": {
                "description": "Cluster-only. Called by scan-service to insert or update assets (subdomains/IPs/ports) discovered during recon phases into the target's inventory.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Upsert discovered assets (INTERNAL)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Assets to upsert",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpsertAssetsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.AssetDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/internal/v1/targets/{target_uid}/verified": {
            "get": {
                "description": "Cluster-only. verified ⇔ target status is \"verified\" AND an active (verified, non-expired) authorization exists AND, when user_id is supplied, the target belongs to that user (reason \"not_owner\" otherwise). Reasons: verified | not_found | not_owner | not_verified | no_active_authorization.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Target verification + ownership check (INTERNAL)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Requesting user's platform id; when set, ownership is enforced",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.VerifiedCheckResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
        "/v1/targets/{target_uid}/assets": {
            "get": {
                "description": "Return the discovered asset inventory (subdomains/IPs/ports/services) for the given target.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assets"
                ],
                "summary": "List assets for a target",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authenticated user ID",
                        "name": "X-User-Id",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Target UID",
                        "name": "target_uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.AssetDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AssetDTO": {
            "type": "object",
            "properties": {
                "asset_type": {
                    "type": "string"
                },
                "discovered_by": {
                    "type": "string"
                },
                "first_seen": {
                    "type": "string"
                },
                "last_seen": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "scope_class": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.AssetUpsertItem": {
            "type": "object",
            "required": [
                "asset_type",
                "value"
            ],
            "properties": {
                "asset_type": {
                    "type": "string",
                    "enum": [
                        "subdomain",
                        "ip",
                        "port",
                        "service",
                        "endpoint"
                    ]
                },
                "discovered_by": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": true
                },
                "value": {
                    "type": "string",
                    "maxLength": 512,
                    "minLength": 1
                }
            }
        },
...
        "model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "description": "\"success\" | \"error\"",
                    "type": "string"
                }
            }
        },
        "model.ScopeCheckRequest": {
            "type": "object",
            "required": [
                "phase",
                "target_uid"
            ],
            "properties": {
                "host": {
                    "description": "Host or IP being queried — exactly one must be provided.",
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "phase": {
                    "type": "string"
                },
                "target_uid": {
                    "type": "string"
                }
            }
        },
        "model.ScopeCheckResponse": {
            "type": "object",
            "properties": {
                "allowed_ips": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "authorized": {
                    "type": "boolean"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "model.UpsertAssetsRequest": {
            "type": "object",
            "required": [
                "assets"
            ],
            "properties": {
                "assets": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.AssetUpsertItem"
                    }
                }
            }
        },
        "model.VerifiedCheckResponse": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                }
            }
        },
...
    }
}